Update for some changes in TMV UI with v0.70 (also still works with v0.65)
git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@7443 2422e80f-9519-0410-a819-8e21e7c86b6d

+++ src/CoaddCatalog.h
@@ -16,7 +16,7 @@ public :
     CoaddCatalog(const ConfigFile& _params);
     ~CoaddCatalog();
 
-    size_t size() const { return _id.size(); }
+    int size() const { return _id.size(); }
 
     void read();
     void readFits(const std::string& file);
